#46 Reorganized accumulo type encoders into mango-core as lexitypeencoders.
