+++ mango-core/src/main/java/org/calrissian/mango/types/encoders/lexi/DateEncoder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.calrissian.mango.accumulo.types.impl;
+package org.calrissian.mango.types.encoders.lexi;
 
 
 import org.calrissian.mango.types.encoders.AbstractDateEncoder;
