+++ mango-json/src/main/java/org/calrissian/mango/json/tuple/TupleModule.java
@@ -20,14 +20,14 @@ import com.fasterxml.jackson.databind.module.SimpleModule;
 import org.calrissian.mango.domain.Tuple;
 import org.calrissian.mango.types.TypeRegistry;
 
-import static org.calrissian.mango.types.GenericTypeEncoders.DEFAULT_TYPES;
+import static org.calrissian.mango.types.SimpleTypeEncoders.SIMPLE_TYPES;
 
 public class TupleModule extends SimpleModule {
 
     private final TypeRegistry<String> typeContext;
 
     public TupleModule() {
-        this(DEFAULT_TYPES);
+        this(SIMPLE_TYPES);
     }
 
     public TupleModule(TypeRegistry<String> typeContext) {
